maintenance: restructured arg options source
