+++ src/fs/autocompletion.ts
@@ -1,5 +1,5 @@
 import { CompleterResult } from "node:readline"
-import { Command, fsCommands, splitCommandSegments } from "../interface/commands"
+import { Command, fsCommands, splitCommandSegments } from "./commands"
 import FilenSDK from "@filen/sdk"
 import { CloudPath } from "../cloudPath"
 import * as fs from "node:fs"
